refactor: split the diesel module and consolidate its tests
Addresses the review feedback on #284:

- Drop the banner comments and section labels that narrated the diff
  rather than documenting the code.
- Split the 1300-line postgis_diesel.rs into reader/sql/writer/tests,
  matching the reader/writer layout used by the geos and geojson
  modules.
- Derive container dimensions from postgis_diesel's own dimension(),
  which already walks to the first point, instead of hand-rolling that
  walk in each of the seven GeozeroGeometry impls.
- Collapse 34 tests that mostly re-tested geozero's WKT and GeoJSON
  writers into 10 covering distinct behaviour: the read path, the write
  path, dimension handling, ring handling, container delegation and
  conversion from a non-WKB source.
