+++ geozero/src/postgis/postgis_diesel/mod.rs
@@ -1,0 +1,22 @@
+//! PostGIS geometry type encoding/decoding for Diesel.
+
+mod reader;
+mod sql;
+#[cfg(test)]
+mod tests;
+mod writer;
+
+pub use writer::PostgisDieselWriter;
+
+pub mod sql_types {
+    use diesel::query_builder::QueryId;
+    use diesel::sql_types::SqlType;
+
+    #[derive(SqlType, QueryId)]
+    #[diesel(postgres_type(name = "geometry"))]
+    pub struct Geometry;
+
+    #[derive(SqlType, QueryId)]
+    #[diesel(postgres_type(name = "geography"))]
+    pub struct Geography;
+}
